Moved Health.cs to scripts folder, fixed indentation
Also it's not in Door because that annoyed me
